+++ libs/utils/release_notes.md
@@ -10,7 +10,7 @@
 2024-03-08
 
 * [new] Extensions/registry; Support extension types filter to get_extensions_from_director, unloads module again if not matching filter.
-* [changed] Extensions/environment; Changed the name of FTRACK_EXTENSIONS_PATH environment variable to FTRACK_CONNECT_EXTENSIONS_PATH.
+* [changed] Extensions/environment; Changed the name of FTRACK_FRAMEWORK_EXTENSIONS_PATH environment variable to FTRACK_CONNECT_EXTENSIONS_PATH.
 * [changed] get_temp_path util now supports filename_extension with or without leading dot(.)